Conflicts:
	lang/ilias_de.lang
	lang/ilias_en.lang

+++ Modules/TestQuestionPool/classes/class.assQuestionGUI.php
@@ -64,6 +64,22 @@ abstract class assQuestionGUI
 	 * @var ilTestQuestionNavigationGUI
 	 */
 	private $navigationGUI;
+
+	const PRESENTATION_CONTEXT_TEST = 'pContextTest';
+	const PRESENTATION_CONTEXT_RESULTS = 'pContextResults';
+
+	/**
+	 * @var string
+	 */
+	private $presentationContext = null;
+
+	const OUTPUT_MODE_SCREEN = 'outModeScreen';
+	const OUTPUT_MODE_PDF = 'outModePdf';
+	
+	/**
+	 * @var string
+	 */
+	private $outputMode = self::OUTPUT_MODE_SCREEN;
 	
 	/**
 	* assQuestionGUI constructor
@@ -71,7 +87,6 @@ abstract class assQuestionGUI
 	function __construct()
 	{
 		global $lng, $tpl, $ilCtrl;
-
 
 		$this->lng =& $lng;
 		$this->tpl =& $tpl;
@@ -130,6 +145,48 @@ abstract class assQuestionGUI
 		return $this->getQuestionType();
 	}
 
+	/**
+	 * @return string
+	 */
+	public function getPresentationContext()
+	{
+		return $this->presentationContext;
+	}
+
+	/**
+	 * @param string $presentationContext
+	 */
+	public function setPresentationContext($presentationContext)
+	{
+		$this->presentationContext = $presentationContext;
+	}
+	
+	public function isTestPresentationContext()
+	{
+		return $this->getPresentationContext() == self::PRESENTATION_CONTEXT_TEST;
+	}
+
+	/**
+	 * @return string
+	 */
+	public function getOutputMode()
+	{
+		return $this->outputMode;
+	}
+
+	/**
+	 * @param string $outputMode
+	 */
+	public function setOutputMode($outputMode)
+	{
+		$this->outputMode = $outputMode;
+	}
+
+	public function isPdfOutputMode()
+	{
+		return $this->getOutputMode() == self::OUTPUT_MODE_PDF;
+	}
+	
 	/**
 	 * @return ilTestQuestionNavigationGUI
 	 */
@@ -899,6 +956,7 @@ abstract class assQuestionGUI
 	{
 	    // title
 		$title = new ilTextInputGUI($this->lng->txt("title"), "title");
+		$title->setMaxLength(100);
 		$title->setValue($this->object->getTitle());
 		$title->setRequired(TRUE);
 		$form->addItem($title);
@@ -1074,6 +1132,11 @@ abstract class assQuestionGUI
 	*/
 	function getAnswerFeedbackOutput($active_id, $pass)
 	{
+		if( $this->isTestPresentationContext() )
+		{
+			return '';
+		}
+		
 		return $this->getGenericFeedbackOutput($active_id, $pass);
 	}
 
@@ -1155,6 +1218,11 @@ abstract class assQuestionGUI
 		
 		return assQuestion::_getQuestionTypeName($this->object->getQuestionType());
 	}
+	
+	public function showSuggestedSolution()
+	{
+		$this->suggestedsolution();
+	}
 
 	/**
 	* Allows to add suggested solutions for questions
@@ -1165,15 +1233,16 @@ abstract class assQuestionGUI
 	{
 		global $ilUser;
 		global $ilAccess;
-		
-		if ($_POST["deleteSuggestedSolution"] == 1)
+
+		$save = (is_array($_POST["cmd"]) && array_key_exists("suggestedsolution", $_POST["cmd"])) ? TRUE : FALSE;
+
+		if ($save && $_POST["deleteSuggestedSolution"] == 1)
 		{
 			$this->object->deleteSuggestedSolutions();
 			ilUtil::sendSuccess($this->lng->txt("msg_obj_modified"), true);
 			$this->ctrl->redirect($this, "suggestedsolution");
 		}
 
-		$save = (is_array($_POST["cmd"]) && array_key_exists("suggestedsolution", $_POST["cmd"])) ? TRUE : FALSE;
 		$output = "";
 		$solution_array = $this->object->getSuggestedSolution(0);
 		$options = array(
@@ -1310,7 +1379,11 @@ abstract class assQuestionGUI
 				$form->addItem($hidden);
 				$form->addItem($question);
 			}
-			if ($ilAccess->checkAccess("write", "", $_GET['ref_id']))	$form->addCommandButton("suggestedsolution", $this->lng->txt("save"));
+			if ($ilAccess->checkAccess("write", "", $_GET['ref_id']))
+			{
+				$form->addCommandButton('showSuggestedSolution', $this->lng->txt('cancel'));
+				$form->addCommandButton('suggestedsolution', $this->lng->txt('save'));
+			}
 			
 			if ($save)
 			{
@@ -1342,6 +1415,7 @@ abstract class assQuestionGUI
 					}
 				}
 			}
+			
 			$output = $form->getHTML();
 		}
 		
@@ -1886,7 +1960,7 @@ abstract class assQuestionGUI
 		{
 			$this->object->setNrOfTries( $_POST['nr_of_tries'] );
 		}
-		$this->object->setQuestion( $_POST['question'] ); // ?
+		$this->object->setQuestion( ilUtil::stripOnlySlashes($_POST['question']) ); // ?
 		$this->object->setEstimatedWorkingTime(
 			$_POST["Estimated"]["hh"],
 			$_POST["Estimated"]["mm"],
